backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# LLM
openai>=1.57.0

# Browser Automation
playwright>=1.49.0
playwright-stealth>=1.0.6

# Resume Parsing
pymupdf>=1.25.0
python-docx>=1.1.0

# Data Models
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Configuration
pyyaml>=6.0.2
python-dotenv>=1.0.1

# CLI
typer>=0.15.0
rich>=13.9.0

# Utilities
httpx>=0.28.0
openpyxl>=3.1.5
